7-Zip 23.01 (x64) runs on DuetOS — region_count widening, CRT thunks, real stdio
Real-world MSVC console PE — 1.29 MiB, 138 imports across KERNEL32 /
msvcrt / ADVAPI32 / OLEAUT32 / USER32 — now loads, runs CRT init,
prints its full ~3 KiB usage block, and exits cleanly (rc=0). 72
lines / 2,905 bytes of help text on serial. Run via:
DUETOS_SMOKE_PROFILE=pe-sevenzip tools/qemu/run.sh

Bugs the bring-up surfaced and fixed (in order they triggered):

1. AddressSpace::region_count was u8 but regions[] holds 1024
   entries. After 256 mappings the counter wrapped, overwriting
   earlier rows. Page tables stayed mapped but
   AddressSpaceLookupUserFrame's linear scan over
   regions[0..count) lost the early entries (the .rdata page
   holding the IAT in particular), and ResolveImports failed
   silently with "IAT slot VA not mapped". Widened to u16 across
   the struct + every loop variable in address_space.cpp /
   panic.cpp / syscall.cpp. Same overflow re-fired in the AS
   destroy loop (FreeFrame on already-free frame at frame
   0xd29000), fixed in the same slice.

2. SpawnPeFile silent-fail on PeLoad rejection: failing spawns
   vanished between the spawn block trace and the next process
   banner. Routed through KLOG_WARN + KLOG_DEBUG_V detail per
   the new "Diagnostic Logging — Keep It, Gate It, Probe It"
   rule.

3. IsLikelyDataImport only recognised C++-mangled names; plain-C
   CRT data exports (_commode/_fmode/_iob/__initenv/_environ/
   __argc/__argv/__wargv/_acmdln/_pgmptr/_amblksiz/__pioinfo/etc.)
   resolved to the function NO-OP stub in the read-only stub
   page, and the CRT init's `mov [iat], value` faulted at
   WRITE_TO_RO_PAGE cr2=0x6000024a. Extended with a 33-name
   plain-C allowlist; future programs that import a CRT global
   we haven't listed surface the same fault path with one-line
   fix.

4. Critical msvcrt CRT functions were unmapped (catch-all
   NO-OP). Without exit, the entry function fell through and
   ret'd to a NULL stack-bottom RIP. Without fputs/fputc/fwrite
   no console output. Wired 12 new thunks_table entries —
   exit/_exit -> kOffExitProcess (NOT kOffTerminate, which forces
   exit code 3 — wrong for arbitrary rc), cleanup-only paths
   (_cexit/_c_exit/_initterm/_onexit/__getmainargs/__set_app_type)
   to kOffReturnZero, fgetc to kOffReturnMinus1.

5. Three new hand-assembled stdio thunks (kOffFputs 34B / kOffFwrite
   28B / kOffFputc 21B) at the end of kThunksBytes, route to
   SYS_WRITE(fd=1, ...). fputc returns c (not 0) because 7za's
   help-print loop checks `if (fputc(...) != c) bail` — a
   ReturnZero stub silently aborted the help text mid-line.
   Updated static_assert(sizeof(kThunksBytes) == 0x109B).

6. kSyscallWriteMax was 256, but 7za prints the full help
   (~3 KiB) as a single fputs call. Bumped to 4096 (one page);
   still bounded, well inside the 16 KiB kernel stack budget.

Smoke profile + ramfs wiring: SmokeProfile::PeSevenZip +
SmokeTarget::PeSevenZip in test/smoke_profile.{h,cpp}, the
"pe-sevenzip" cmdline parser case, gated to !emulator under
profile=None (the load + 138 imports cost ~10 s extra TCG wall),
30-second sleep tick budget. Spawn site mirrors ring3-winkill.
PE embedded as /bin/7za.exe in ramfs via kBinSevenZipBytes.

Verification — all four smoke profiles complete + sentinel +
zero bare FAIL + session OK after the changes:
- bringup ✓, pe-winapi ✓, pe-sevenzip ✓ (72-line help printed),
  linux ✓.

+++ kernel/core/panic.cpp
@@ -418,7 +418,7 @@ void DumpProcessVmInfo()
     {
         u64 vmin = ~static_cast<u64>(0);
         u64 vmax = 0;
-        for (u8 i = 0; i < as->region_count; ++i)
+        for (u16 i = 0; i < as->region_count; ++i)
         {
             const u64 v = as->regions[i].vaddr;
             if (v < vmin)
